refactor: move prohibited user names to modules
In anticipation of adding more routes and tests.
