+++ tests/unit/admin/views/test_users.py
@@ -24,7 +24,6 @@ from warehouse.accounts.interfaces import IEmailBreachedService, IUserService
 from warehouse.accounts.models import (
     DisableReason,
     ProhibitedEmailDomain,
-    ProhibitedUserName,
     RecoveryCode,
     WebAuthn,
 )
@@ -1472,60 +1471,3 @@ class TestUserRecoverAccountComplete:
         assert existing_primary_email.verified is True
         assert user.primary_email == override_to_email
         assert user.primary_email.verified is True
-
-
-class TestBulkAddProhibitedUserName:
-    def test_get(self):
-        request = pretend.stub(method="GET")
-
-        assert views.bulk_add_prohibited_user_names(request) == {}
-
-    def test_bulk_add(self, db_request):
-        db_request.user = UserFactory.create()
-        db_request.method = "POST"
-
-        already_existing_prohibition = ProhibitedUserName(
-            name="prohibition-already-exists",
-            prohibited_by=db_request.user,
-            comment="comment",
-        )
-        db_request.db.add(already_existing_prohibition)
-
-        already_existing_user = UserFactory.create(username="user-already-exists")
-        UserFactory.create(username="deleted-user")
-
-        user_names = [
-            already_existing_prohibition.name,
-            already_existing_user.username,
-            "doesnt-already-exist",
-        ]
-
-        db_request.POST["users"] = "\n".join(user_names)
-
-        db_request.session = pretend.stub(
-            flash=pretend.call_recorder(lambda *a, **kw: None)
-        )
-        db_request.route_path = lambda a: "/admin/prohibited_user_names/bulk"
-
-        result = views.bulk_add_prohibited_user_names(db_request)
-
-        assert db_request.session.flash.calls == [
-            pretend.call(
-                f"Prohibited {len(user_names)!r} users",
-                queue="success",
-            )
-        ]
-        assert result.status_code == 303
-        assert result.headers["Location"] == "/admin/prohibited_user_names/bulk"
-
-        for user_name in user_names:
-            prohibition = (
-                db_request.db.query(ProhibitedUserName)
-                .filter(ProhibitedUserName.name == user_name)
-                .one()
-            )
-
-            assert prohibition.name == user_name
-            assert prohibition.prohibited_by == db_request.user
-
-            assert db_request.db.query(User).filter(User.name == user_name).count() == 0
